+++ docs/main/changelog.rst
@@ -9,6 +9,7 @@ follows `Semantic versioning`_
 
 Development version
 -------------------
+- Add ``Configuration.from_yaml()`` method to load configuration from yaml file.
 - Add ``Configuration.from_ini()`` method to load configuration from ini file.
 - Add ``Configuration.from_dict()`` method to load configuration from dictionary.
 - Remove undocumented positional parameter of ``DependenciesContainer`` provider.